Refactor app script into modular architecture
